TowerShooter v1.02
Cleaned up some code, removed the PosTower from the screen and replaced
with the actual tank. Removed the tank from a group since it is the only
tank you control. Added "Levels", each level consists of 5 waves, each
wave adds an enemy, when you go on to next level, the position and
number of trees change, the enemies on screen get reset to the original
amount, and their health goes up by 1. If a base gets destroyed it
permanently adds an enemy to the original amount.