TowerShooter v1.03
Created an enemy vehicle that moves faster than the enemies and appear
as sort of a boss at the end of wave 5, more are added the higher the
level you go. Fixed the problem when you finished the game that sent you
back to the splash screen. The powerups now have a time limit before the
disappear. Changed the instructions font colour so you can read it now.